feat(lib): add thumb-hash lib and change path alias
