+++ apps/web/app/(main)/albums/components/delete-album-modal.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/navigation';
 
-import { request } from '@/utils/request';
+import { request } from '~/utils/request';
 
 interface DeleteAlbumModalProps {
   album: Album;
