+++ apps/web/app/(main)/explore/page.tsx
@@ -1,36 +1,8 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-
-import PageHeader from '@/components/page-header';
-import { request } from '@/utils/request';
-
-interface Settings {
-  photoDirs: string[];
-}
+import PageHeader from '~/components/page-header';
 
 export default function Page() {
-  const [settings, setSettings] = useState<Settings>({ photoDirs: [] });
-
-  useEffect(() => {
-    (async () => {
-      const settings = await request('/api/v1/settings');
-      settings && setSettings(settings.data);
-    })();
-  }, []);
-
-  const updateSettings = async ({ photoDirs }: Settings) => {
-    const newSettings = await request('/api/v1/settings', {
-      method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ photoDirs }),
-    });
-
-    newSettings && setSettings(newSettings.data);
-  };
-
   return (
     <div className="absolute inset-0 overflow-y-auto overflow-x-hidden transition-all duration-500">
       <PageHeader title="Explore"></PageHeader>
