+++ apps/web/app/(main)/photos/page.tsx
@@ -3,17 +3,16 @@
 import clsx from 'clsx';
 import { useEffect, useState } from 'react';
 
-import PageHeader from '@/components/page-header';
-import Photos from '@/components/photos';
-import PhotosLayoutSetting from '@/components/photos/photos-layout-setting';
-import IconButton from '@/components/ui/icon-button';
-import Upload from '@/components/upload';
-import { DEFAULT_PHOTOS_LAYOUT } from '@/config/constants';
-import { NavMode } from '@/config/enums';
-import FilterAltIcon from '@/icons/filter-alt-icon';
-import SelfImprovementIcon from '@/icons/self-improvement-icon';
-import TuneIcon from '@/icons/tune-icon';
-import { request } from '@/utils/request';
+import PageHeader from '~/components/page-header';
+import Photos from '~/components/photos';
+import PhotosLayoutSetting from '~/components/photos/photos-layout-setting';
+import IconButton from '~/components/ui/icon-button';
+import Upload from '~/components/upload';
+import { DEFAULT_PHOTOS_LAYOUT } from '~/config/constants';
+import { NavMode } from '~/config/enums';
+import FilterAltIcon from '~/icons/filter-alt-icon';
+import TuneIcon from '~/icons/tune-icon';
+import { request } from '~/utils/request';
 
 import { useNavMode } from '../nav-provider';
 import { groupPhotosByDate } from './utils';
