+++ libs/thumb-hash/README.md
@@ -1,0 +1,7 @@
+# thumb-hash
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test thumb-hash` to execute the unit tests via [Jest](https://jestjs.io).
