+++ libs/thumb-hash/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/thumb-hash';
