+++ tsconfig.base.json
@@ -14,7 +14,9 @@
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
-    "paths": {}
+    "paths": {
+      "@here-photos/thumb-hash": ["libs/thumb-hash/src/index.ts"]
+    }
   },
   "exclude": ["node_modules", "tmp"]
 }
